refactoring IntCodeComputer for Day 9

+++ AOC-2019/src/main/java/net/thebyrdnest/aoc/aoc2019/Day02.java
@@ -3,7 +3,7 @@ package net.thebyrdnest.aoc.aoc2019;
 import net.thebyrdnest.aoc.utils.IntCodeComputer;
 
 public class Day02 {
-    public int solve1(int[] mem) {
+    public long solve1(long[] mem) {
         IntCodeComputer intCodeComputer = new IntCodeComputer(1, mem);
         intCodeComputer.run();
         while (!intCodeComputer.isDone()) {
@@ -17,7 +17,7 @@ public class Day02 {
         return intCodeComputer.getMemoryValue(0);
     }
 
-    public int solve2(int[] mem, int target) {
+    public int solve2(long[] mem, long target) {
         boolean bDone = false;
         for (int noun = 0; noun <= 99; noun++) {
             for (int verb = 0; verb <= 99; verb++) {
